Removing CrabZ80's core. (it was never working and i doubt it would be accurate anyway.)
Updating the emulator's core code to fix some issues reported by Clang/GCC.
(bios.pages was 8-bits even though cart.pages was 16-bits.
This could cause issues if a game like Bad Apple is played using the bios.)

Removing error.c/error.h, it was never used.
I'm trying to fix the signing/unsigned issues that i had introduced
a while back unfortunately (though the older code still suffered from it
to some code).

Some of it is fixed but not all of it is fixed. Ideally, this should
be switched to using 32-bits ints if possible.