Fixes for Gauntley display issues.
There were two issues with that game :
1) It tries to display a sprite being overbound.
The simple fix for this is to use an unsigned 8-bits variable rather
than using a signed 32-bits value and checking if it goes overbound.
This simplification fixes the sprite issue in the upper right of the screen
when ingame.

2) It would not scroll the HUD properly.
Not quite sure what's the issue exactly but it seems that
it shouldn't use v_line but line instead when stopping the vscroll.

This fixes the vertical scrolling in Gauntlet.

All of these issues were introduced by ekeeke when he revamped the code
in SMS Plus GX.