Add FBDev/evdev port for Embedded linux platforms.
This was mostly done to allow us to decrease the size of the binary
when using static linking as well as avoid the overhead of libSDL.

The input code reads directly from /dev/input/event0 (which is the keyboard).
It does seem to suffer from ghosting (?) when the program slows down a lot.
This does not happen on libSDL so i'm not sure what to do,
it seems that they might be using threads to solve this issue ?

The FBDev code does not currently support double buffering :
This would require us doubling the vertical size of the allocated
resolution.
The Pandora wiki explains how this is implemented with fbdev :
https://pandorawiki.org/Kernel_interface